split range
