+++ src/range/Range1D.ts
@@ -1,0 +1,508 @@
+/**
+ * Created by Samuel Gratzl on 27.12.2016.
+ */
+
+import {IRangeElem} from './internal';
+import RangeElem from './internal/RangeElem';
+import {IIterator, Iterator, forList, concat} from '../iterator';
+import CompositeRange1D from './CompositeRange1D';
+
+export default class Range1D {
+  private readonly arr: IRangeElem[];
+
+  constructor(arg?: Range1D|IRangeElem[]) {
+    if (arg instanceof Range1D) {
+      this.arr = (<Range1D>arg).arr;
+    } else if (Array.isArray(arg)) {
+      this.arr = <IRangeElem[]>arg;
+    } else {
+      this.arr = [];
+    }
+  }
+
+  get length() {
+    return this.size();
+  }
+
+  static all() {
+    return new Range1D([RangeElem.all()]);
+  }
+
+  static single(item: number) {
+    return new Range1D([RangeElem.single(item)]);
+  }
+
+  static none() {
+    return new Range1D();
+  }
+
+  static from(indices: number[]) {
+    return new Range1D(Range1D.compress(indices));
+  }
+
+  private static compress(indices: number[]) {
+    if (indices.length === 0) {
+      return [];
+    } else if (indices.length === 1) {
+      return [RangeElem.single(indices[0])];
+    }
+    //return indices.map(RangeElem.single);
+    const r = <IRangeElem[]>[],
+      deltas = indices.slice(1).map((e, i) => e - indices[i]);
+    let start = 0, act = 1, i = 0;
+    while (act < indices.length) {
+      while (deltas[start] === deltas[act - 1] && act < indices.length) { //while the same delta
+        act++;
+      }
+      if (act === start + 1) { //just a single item used
+        r.push(RangeElem.single(indices[start]));
+      } else {
+        //+1 since end is excluded
+        //fix while just +1 -1 is allowed
+        if (Math.abs(deltas[start]) === 1) {
+          r.push(RangeElem.range(indices[start], indices[act - 1] + deltas[start], deltas[start]));
+        } else {
+          for (i = start; i < act; i++) {
+            r.push(RangeElem.single(indices[i]));
+          }
+        }
+      }
+      start = act;
+      act += 1;
+    }
+    while (start < indices.length) { //corner case by adding act+1, it might happen that the last one isnt considered
+      r.push(RangeElem.single(indices[start++]));
+    }
+    return r;
+  }
+
+  get isAll() {
+    return this.arr.length === 1 && this.at(0).isAll;
+  }
+
+  get isNone() {
+    return this.arr.length === 0;
+  }
+
+  get isUnbound() {
+    return this.arr.some((d) => d.isUnbound);
+  }
+
+  private get isList() {
+    return this.arr.every((d) => d.isSingle);
+  }
+
+  push(...items: (IRangeElem|string|number|[number, number, number])[]): number {
+    function p(item: any) {
+      if (typeof item === 'string') {
+        return RangeElem.parse(item.toString());
+      } else if (typeof item === 'number') {
+        return RangeElem.single(<number>item);
+      } else if (Array.isArray(item)) {
+        return new RangeElem(item[0], item[1], item[2]);
+      }
+      return <RangeElem>item;
+    }
+
+    return this.arr.push(...items.map(p));
+  }
+
+  pushSlice(from: number, to: number = -1, step: number = 1) {
+    this.arr.push(new RangeElem(from, to, step));
+  }
+
+  pushList(indices: number[]) {
+    this.arr.push(...Range1D.compress(indices));
+  }
+
+  setSlice(from: number, to: number = -1, step: number = 1) {
+    this.arr.length = 0;
+    this.pushSlice(from, to, step);
+  }
+
+  setList(indices: number[]) {
+    this.arr.length = 0;
+    this.pushList(indices);
+  }
+
+  at(index: number): IRangeElem {
+    if (index < 0) {
+      index += this.length;
+    }
+    if (index < 0 || index >= this.arr.length) {
+      return RangeElem.none();
+    }
+    return this.arr[index];
+  }
+
+  size(size?: number) {
+    const t = this.arr.map((d) => d.size(size));
+    return t.reduce((a, b) => a + b, 0);
+  }
+
+  /**
+   * whether this range is the identity, i.e. the first natural numbers starting with 0
+   * @return {boolean}
+   */
+  get isIdentityRange() {
+    return this.arr.length === 1 && this.arr[0].from === 0 && this.arr[0].step === 1;
+  }
+
+  repeat(ntimes = 1) {
+    if (ntimes === 1) {
+      return this;
+    }
+    const r = this.arr.slice();
+    //push n times
+    for (let i = 1; i < ntimes; ++i) {
+      r.push.apply(r, this.arr);
+    }
+    return new Range1D(r);
+  }
+
+  /**
+   * combines this range with another and returns a new one
+   * this = (1,3,5,7), sub = (1,2) -> (1,2)(1,3,5,7) = (3,5)
+   * @returns {*}
+   */
+  preMultiply(sub: Range1D, size?: number): Range1D {
+    if (this.isAll) {
+      return sub.clone();
+    }
+    if (sub.isAll) {
+      return this.clone();
+    }
+    if (sub.isNone || this.isNone) {
+      return Range1D.none();
+    }
+    if (this.isIdentityRange) { //identity lookup
+      return sub.clone();
+    }
+    //TODO optimize
+    const l = this.iter(size).asList();
+    const mapImpl = (sub) => {
+      const s = sub.iter(l.length);
+      const r = [];
+      s.forEach((i) => {
+        if (i >= 0 && i < l.length) { //check for out of range
+          r.push(l[i]);
+        }
+      });
+      return sub.fromLike(r);
+    };
+
+    if (sub instanceof CompositeRange1D) {
+      return new CompositeRange1D(sub.name, sub.groups.map(mapImpl));
+    } else {
+      return mapImpl(sub);
+    }
+  }
+
+  /**
+   * logical union between two ranges
+   * @param other
+   * @returns {RangeDim}
+   */
+  union(other: Range1D, size?: number): Range1D {
+    if (this.isAll || other.isNone) {
+      return this.clone();
+    }
+    if (other.isAll || this.isNone) {
+      return other.clone();
+    }
+    const r = this.iter(size).asList();
+    const it2 = other.iter(size);
+    it2.forEach((i) => {
+      if (r.indexOf(i) < 0) {
+        r.push(i);
+      }
+    });
+    return other.fromLike(r.sort());
+  }
+
+  /**
+   * logical intersection between two ranges
+   * @param other
+   * @returns {RangeDim}
+   */
+  intersect(other: Range1D, size?: number) {
+    if (this.isNone || other.isNone) {
+      return Range1D.none();
+    }
+    if (this.isAll) {
+      return other.clone();
+    }
+    if (other.isAll) {
+      return this.clone();
+    }
+    const it1 = this.iter(size).asList();
+    const it2 = other.iter(size);
+    const r = [];
+    it2.forEach((i) => {
+      if (it1.indexOf(i) >= 0) {
+        r.push(i);
+      }
+    });
+    return other.fromLike(r.sort());
+  }
+
+  toSet(size?: number) {
+    return this.removeDuplicates(size);
+  }
+
+  /**
+   * logical difference between two ranges
+   * @param without
+   * @returns {RangeDim}
+   */
+  without(without: Range1D, size?: number) {
+    if (this.isNone || without.isNone) {
+      return this.clone();
+    }
+    if (without.isAll) {
+      return Range1D.none();
+    }
+    const it1 = this.iter(size);
+    const it2 = without.iter(size).asList();
+    const r = [];
+    it1.forEach((i) => {
+      if (it2.indexOf(i) < 0) {
+        r.push(i);
+      }
+    });
+    return Range1D.from(r.sort());
+  }
+
+  /**
+   * clones this range
+   * @returns {RangeDim}
+   */
+  clone() {
+    return new Range1D(this.arr.map((d) => d.clone()));
+  }
+
+  /**
+   * inverts the given index to the original range
+   * @param index
+   * @param size the underlying size for negative indices
+   * @returns {*}
+   */
+  invert(index: number, size?: number) {
+    if (this.isAll) {
+      return index;
+    }
+    if (this.isNone) {
+      return -1; //not mapped
+    }
+    let s = this.arr[0].size(size);
+    let act = 0, total: number = s;
+    while (total > index && act < this.length) {
+      act++;
+      s = this.arr[act].size(size);
+      total += s;
+    }
+    if (act >= this.arr.length) {
+      return -1; //not mapped
+    }
+    return this.arr[act - 1].invert(index - total + s, size);
+  }
+
+  indexOf(indices: number[]): number[];
+
+  indexOf(index: number): number;
+
+  indexOf(...index: number[]): number[];
+
+  indexOf(r: Range1D, size?: number): Range1D;
+
+  /**
+   * returns the index(ices) of the given elements
+   * @return {*}
+   */
+  indexOf(): any {
+    if (arguments[0] instanceof Range) {
+      return this.indexRangeOf(arguments[0], arguments[1]);
+    }
+    let arr: number[];
+    const base = this.iter().asList();
+    if (arguments.length === 1) {
+      if (typeof arguments[0] === 'number') {
+        return base.indexOf(<number>arguments[0]);
+      }
+      arr = arguments[0];
+    } else {
+      arr = Array.from(arguments);
+    }
+    if (arr.length === 0) {
+      return [];
+    }
+    return arr.map((index, i) => base.indexOf(index));
+  }
+
+  /**
+   * returns the range representing the indices of the given range within the current data
+   * @param r
+   * @param size
+   * @return {Range1D}
+   */
+  indexRangeOf(r: Range1D, size?: number): Range1D {
+    if (r.isNone || this.isNone) {
+      return r.fromLike([]);
+    }
+    if (r.isAll) { //index of all is still all
+      return Range1D.all();
+    }
+    //
+    let mapImpl;
+    if (this.isIdentityRange) {
+      let end = this.arr[0].to;
+      mapImpl = (d, result) => {
+        if (d >= 0 && d < end) {
+          result.push(d);
+        }
+      };
+    } else {
+      const arr = this.iter().asList();
+      mapImpl = (d, result) => {
+        const i = arr.indexOf(d);
+        if (i >= 0) {
+          result.push(i);
+        }
+      };
+    }
+    if (r instanceof CompositeRange1D) {
+      return new CompositeRange1D(r.name, r.groups.map((g) => {
+        const result = [];
+        g.forEach((d) => mapImpl(d, result));
+        return g.fromLike(result);
+      }));
+    } else {
+      const result = [];
+      r.forEach((d) => mapImpl(d, result));
+      return r.fromLike(result);
+    }
+  }
+
+  /**
+   * filters the given data according to this range
+   * @param data
+   * @param size the total size for resolving negative indices
+   * @returns {*}
+   */
+  filter(data: any[], size?: number, transform: (any) => any = (d) => d) {
+    if (this.isAll) {
+      return data.map(transform);
+    }
+    const it = this.iter(size);
+    //optimization
+    if (it.byOne && it instanceof Iterator) {
+      return data.slice((<Iterator><any>it).from, (<Iterator><any>it).to).map(transform);
+      //} else if (it.byMinusOne) {
+      //  var d = data.slice();
+      //  d.reverse();
+      //  return d;
+    } else {
+      const r = [];
+      while (it.hasNext()) {
+        r.push(transform(data[it.next()]));
+      }
+      return r;
+    }
+  }
+
+  /**
+   * creates an iterator of this range
+   * @param size the underlying size for negative indices
+   */
+  iter(size?: number): IIterator<number> {
+    if (this.isList) {
+      return forList(this.arr.map((d) => (<any>d).from));
+    }
+    const its: IIterator<number>[] = this.arr.map((d) => d.iter(size));
+    return concat.apply(null, its);
+  }
+
+  get __iterator__() {
+    return this.iter();
+  }
+
+  asList(size?: number): number[] {
+    return this.iter(size).asList();
+  }
+
+  get first() {
+    if (this.isNone) {
+      return null;
+    }
+    return this.arr[0].from;
+  }
+
+  get last() {
+    if (this.isNone) {
+      return null;
+    }
+    return this.arr[this.arr.length - 1].from;
+  }
+
+  /**
+   * for each element
+   * @param callbackfn
+   * @param thisArg
+   */
+  forEach(callbackfn: (value: number, index: number) => void, thisArg?: any): void {
+    return this.iter().forEach(callbackfn, thisArg);
+  }
+
+  contains(value: number, size?: number): boolean {
+    return this.arr.some((elem) => elem.contains(value, size));
+  }
+
+  /**
+   * sort
+   * @param cmp
+   * @return {Range1D}
+   */
+  sort(cmp: (a: number, b: number) => number = (a, b) => a - b): Range1D {
+    const arr = this.iter().asList();
+    const r = arr.sort(cmp);
+    return this.fromLike(r);
+  }
+
+  private removeDuplicates(size?: number): Range1D {
+    let arr = this.iter().asList();
+    arr = arr.sort();
+    arr = arr.filter((di, i) => di !== arr[i - 1]); //same value as before, remove
+    return Range1D.from(arr);
+  }
+
+  /**
+   * reverts the order of this range
+   */
+  reverse(): Range1D {
+    const a = this.arr.map((r) => r.reverse());
+    a.reverse();
+    return new Range1D(a);
+  }
+
+  toString() {
+    if (this.isAll) {
+      return '';
+    }
+    if (this.length === 1) {
+      return this.arr[0].toString();
+    }
+    return '(' + this.arr.join(',') + ')';
+  }
+
+  eq(other: Range1D) {
+    if (this === other || (this.isAll && other.isAll) || (this.isNone && other.isNone)) {
+      return true;
+    }
+    //TODO more performant comparison
+    return this.toString() === other.toString();
+  }
+
+  fromLike(indices: number[]) {
+    return Range1D.from(indices);
+  }
+}
